==> zindi_challenge/1.Add_variable.ipynb <==
# Zindi - Challenge

https://zindi.africa/competitions/predict-the-global-spread-of-covid-19

April 9, 2020

import pandas as pd
import seaborn as sns
import matplotlib as plt 
import requests
import lxml.html as lh

Only the folder Global_variable_per-country was used.

## Import Data

#Read train 
deaths = pd.read_csv('data/deaths.csv')
cases = pd.read_csv('data/cases.csv')
deaths.drop(['Province/State'], axis=True)
deaths.shape

#transpose data frame
deaths.T 

deaths.head(2)

Compare the 1st 3 columns.

#no missing values
sum(deaths['Country/Region'].isnull() == False)

l1=deaths['Country/Region'].unique()
l2=deaths['Territory'].unique()
len([value for value in l1 if value in l2] )

len([x for x in l2 if x not in l1])
[x for x in l2 if x not in l1]

len([x for x in l1 if x not in l2])
[x for x in l1 if x not in l2]

len(deaths['Territory'].unique())

len(deaths['Province/State'].unique())

deaths[deaths['Province/State'].isnull() == False]

We will use 'Country/Region'. 


country = deaths['Country/Region']

deaths.iloc[220:225,]

deaths[deaths.Population.isnull()]

## Fill in NaN values for Population
index = deaths[deaths.Population.isnull()].index
# updated in 2017
deaths.at[index[0], 'Population'] = 76965 
deaths.at[index[1], 'Population'] = 395361
deaths.at[index[2], 'Population'] = 11.05e+06
deaths.at[index[3], 'Population'] = 428697
deaths.at[index[4], 'Population'] = 546388 
deaths.at[index[5], 'Population'] = 4.659e+06
deaths.at[index[6], 'Population'] = 5.261e+06
deaths.at[index[7], 'Population'] = 81.34e+06
deaths.at[index[8], 'Population'] = 24.29e+06 # Cote d'Ivoire
deaths.at[index[10], 'Population'] = 10.77e+06
deaths.at[index[11], 'Population'] = 1.367e+06
deaths.at[index[12], 'Population'] = 2.101e+06
deaths.at[index[13], 'Population'] = 81.16e+06
deaths.at[index[14], 'Population'] = 51.47e+06
deaths.at[index[15], 'Population'] = 3.55e+06
deaths.at[index[17], 'Population'] = 105264
deaths.at[index[18], 'Population'] = 161014
deaths.at[index[19], 'Population'] = 41109 
deaths.at[index[21], 'Population'] = 21.48e+06
deaths.at[index[23], 'Population'] = 104.9e+06
deaths.at[index[24], 'Population'] = 144.5e+06
deaths.at[index[25], 'Population'] = 7.022e+06
deaths.at[index[26], 'Population'] = 12.58e+06
deaths.at[index[28], 'Population'] = 57.31e+06
deaths.at[index[29], 'Population'] = 9.4e+06
deaths.at[index[30], 'Population'] = 65441
deaths.at[index[31], 'Population'] = 61559
deaths.at[index[32], 'Population'] = 170499
deaths.at[index[33], 'Population'] = 34571
deaths.at[index[34], 'Population'] = 84287
deaths.at[index[35], 'Population'] = 5900 #2008
deaths.at[index[38], 'Population'] = 31.98e+06
deaths.at[index[39], 'Population'] = 95.54e+06
deaths.at[index[40], 'Population'] = 18.27e+06
deaths.at[index[41], 'Population'] = 1.296e+06
deaths.at[index[42], 'Population'] = 6.858e+06
deaths.at[index[43], 'Population'] = 15094
deaths.at[index[44], 'Population'] = 31196
deaths.at[index[45], 'Population'] = 35446
deaths.at[index[46], 'Population'] = 25157
deaths.at[index[47], 'Population'] = 2840
deaths.at[index[48], 'Population'] = 204327
deaths.at[index[49], 'Population'] = 813912
deaths.at[index[50], 'Population'] = 53127
deaths.at[index[51], 'Population'] = 105544
deaths.at[index[52], 'Population'] = 53.37e+06
deaths.at[index[53], 'Population'] = 51.47e+06
deaths.at[index[54], 'Population'] = 40.53e+06

#2018 
deaths.at[index[27], 'Population'] = 23.78e+06
deaths.at[index[37], 'Population'] = 327.2e+06
#2019
deaths.at[index[9], 'Population'] = 10.65e+06
deaths.at[index[20], 'Population'] = 17.28e+06
deaths.at[index[22], 'Population'] = 2.077e+06
deaths.at[index[36], 'Population'] = 6.65e+06

#2020
deaths.at[index[16], 'Population'] = 631219





No more NaN values for population. 

Add a code country column.

def web_scraping_data(link):
    #Create a handle, page, to handle the contents of the website
    page = requests.get(link)
    doc = lh.fromstring(page.content)
    tr_elements = doc.xpath('//tr')

    tr_elements = doc.xpath('//tr')
    col=[]
    i=0
    for t in tr_elements[0]:
        i+=1
        name=t.text_content()
        #print ('%d:"%s"'%(i,name))
        col.append((name,[]))

    for j in range(1,len(tr_elements)):
        T=tr_elements[j]
        i=0

        for t in T.iterchildren():
            data=t.text_content() 
            if i>0:
                try:
                    data=int(data)
                except:
                    pass

            col[i][1].append(data)
            i+=1

    Dict={title:column for (title,column) in col}
    df=pd.DataFrame(Dict)
    
    return df

code_country = web_scraping_data("https://www.iban.com/country-codes")
df = deaths.merge(code_country, how='left', left_on='Country/Region', right_on='Country')

df[df['Alpha-3 code'].isnull()]

## Fill in NaN values for alpha 3 code 
index = df[df['Alpha-3 code'].isnull()].index
variable = 'Alpha-3 code'
df.at[index[0], variable] = 'BHS'
df.at[index[1], variable] = 'BOL'
df.at[index[2], variable] = 'BRN'
df.at[index[3], variable] = 'CAF'
df.at[index[4], variable] = 'COG'
df.at[index[5], variable] = 'COD'
df.at[index[6], variable] = 'CIV'
df.at[index[7], variable] = 'DOM'
df.at[index[8], variable] = 'GMB'
df.at[index[9],variable] = 'IRN'
df.at[index[10],variable] = 'PRK'
df.at[index[11], variable] = 'MDA'
df.at[index[12], variable] = 'ABW'
df.at[index[13], variable] = 'CUW'
df.at[index[14],variable] = 'SXM'
df.at[index[15],variable] = 'NLD'
df.at[index[16],variable] = 'NER'
df.at[index[17], variable] = 'MKD'
df.at[index[18], variable] = 'PHL'
df.at[index[19], variable] = 'RUS'
df.at[index[20], variable] = 'SSD'
df.at[index[21], variable] = 'TWN'
df.at[index[22],variable] = 'TZA'
df.at[index[23],variable] = 'ARE'
df.at[index[24],variable] = 'BMU'
df.at[index[25],variable] = 'CYM'
df.at[index[26],variable] = 'GBR' #by default
df.at[index[27],variable] = 'GIB'
df.at[index[28],variable] = 'IMN'
df.at[index[29],variable] = 'MSR'
df.at[index[30],variable] = 'GBR'
df.at[index[31],variable] = 'USA'
df.at[index[32], variable] = 'VEN'
df.at[index[33],variable] = 'VNM'
df.at[index[34],variable] = 'SYR'
df.at[index[35], variable] = 'LAO'
df.at[index[36], variable] = 'AIA'
df.at[index[37], variable] = 'VGB'
df.at[index[38], variable] = 'TCA'
df.at[index[39],variable] = 'BES'
df.at[index[40], variable] = 'FLK'
df.at[index[41], variable] = 'KOR'


We will use alpha 3 code to join with other tables.

## Add variables


### Temperature 

Source 
- https://datacatalog.worldbank.org/dataset/climate-change-knowledge-portal-historical-data

temp = pd.read_csv('variable/Global_variable_per-country/temp.csv')

temp.head()

# temperature = pd.merge(left=temp, right=df, left_on='ISO_3DIGIT', right_on='Alpha-3 code')

df = df.merge(temp, how='left', left_on='Alpha-3 code', right_on='ISO_3DIGIT')
df.drop(['Numeric','ISO_3DIGIT'], axis =1, inplace=True)

sum(df.Annual_temp.isnull())

### Other epidemy 

SARS-CoV

China Canada, Hong Kong Special Administrative Region of China, Chinese Taipei, Singapore, and Hanoi in Viet Nam.

Ebola 

The Democratic Republic of the Congo
Gabon
Guinea
Italy
Ivory Coast
Liberia
Mali
Nigeria
Philippines
Russia
Senegal
Sierra Leone
South Africa
South Sudan
Spain
Uganda
United Kingdom
United States
What causes Ebol

### Mers

https://www.kaggle.com/imdevskp/mers-outbreak-dataset-20122019

#nb of case in each country for MERS
mers = pd.read_csv('variable/Global_variable_per-country/mers/country_count_latest.csv')
# mers = pd.read_csv('variable/mers/weekly_clean.csv') # to follow week per week 
mers.shape

df = df.merge(mers, how='left', left_on='Territory', right_on='Country')
df['Mers-confirmed'] = df.Confirmed 
df.drop(['Confirmed'], axis=1, inplace=True)

df["Mers-confirmed"] = df["Mers-confirmed"].fillna(0)

### Sars 
https://www.kaggle.com/imdevskp/sars-outbreak-2003-complete-dataset

sars = pd.read_csv('variable/Global_variable_per-country/sars/sars_2003.csv')
c = list(sars.Country.unique())
death_nb = []
for country in c:
    v = list(sars[sars.Country == country]['Number of deaths'])[len(sars[sars.Country ==country]['Number of deaths'])-1]
    death_nb.append(v)

sars = pd.DataFrame({'Country':c, 'death_nb_sars': death_nb})
df = df.merge(sars, how='left', left_on='Territory', right_on='Country')
df.drop(['Country_y',  'Country'], axis=1, inplace=True)
df["death_nb_sars"] = df["death_nb_sars"].fillna(0)

### Age / density / wealth /sex 

See relevant indicators

http://datatopics.worldbank.org/universal-health-coverage/coronavirus/



https://datacatalog.worldbank.org/search?search_api_views_fulltext_op=AND&f%5B0%5D=field_collection_field%3A2026&sort_by=field_wbddh_modified_date

https://data.worldbank.org/indicator/en.pop.dnst

def wordbankdata(variable, tab):
    data = pd.read_csv('variable/Global_variable_per-country/'+variable+'.csv')
    data = data[['Country Code', '2018']]
    tab = df.merge(data, how='left', left_on='Alpha-3 code', right_on='Country Code')
    tab[variable] = tab['2018']
    tab.drop(['2018', 'Country Code'], axis=True, inplace=True)
    return(tab)

variable = '%+65'

df = wordbankdata(variable, df)

variable = '15-64pop'
df = wordbankdata(variable, df)
df = wordbankdata('density', df)
df = wordbankdata( 'sex', df)

## not so many NaN cool !
print(sum(df['%+65'].isnull()))
print(sum(df['15-64pop'].isnull()))
print(sum(df['density'].isnull()))
print(sum(df['sex'].isnull()))

rich = pd.read_csv('variable/Global_variable_per-country/income.csv')
rich = rich[['Country Code', 'IncomeGroup']]
df = df.merge(rich, how='left', left_on='Alpha-3 code', right_on='Country Code')
df.drop(['Country Code'], axis=1, inplace=True)

print(sum(df['IncomeGroup'].isnull()))

## Tourism

tourism = pd.read_csv('variable/Global_variable_per-country/tourism.csv')

tourism.count() / len(tourism)
Lets take 2011.

tourism = tourism[['2011','Country Code']]
df = df.merge(tourism, how='left', left_on='Alpha-3 code', right_on='Country Code')
df.drop(['Country Code'], axis=1, inplace=True)

df['tourism'] = df['2011']
df.drop(['2011'], axis=1, inplace=True)

## Economical factors

https://www.kaggle.com/lewisduncan93/the-economic-freedom-index

econ = pd.read_csv('variable/Global_variable_per-country/economic_freedom_index2019_data.csv',encoding='latin-1')

econ.drop(['Region Rank','Tax Burden % of GDP','Trade Freedom','Financial Freedom','Property Rights','FDI Inflow (Millions)','Judical Effectiveness','Business Freedom','Tax Burden','Labor Freedom','Population (Millions)','Monetary Freedom'], axis=1, inplace=True)

df = df.merge(econ, how='left', left_on='Country/Region', right_on='Country')

df.drop(['Country'], axis=1, inplace=True)

## Immunization factor

https://www.kaggle.com/lsind18/who-immunization-coverage

BCG.csv: the percentage of 1-year-olds who have received one dose of bacille Calmette-Guérin (BCG) vaccine in a given year.

DTP3.csv: the percentage of 1-year-olds who have received three doses of the combined diphtheria, tetanus toxoid and pertussis vaccine in a given year.

HepB3.csv: the percentage of 1-year-olds who have received three doses of hepatitis B vaccine in a given year.

Hib3.csv: the percentage of 1-year-olds who have received three doses of Haemophilus influenzae type B vaccine in a given year.

MCV1.csv: the percentage of children under 1 year of age who have received at least one dose of measles-containing vaccine in a given year. For countries recommending the first dose of measles vaccine in children over 12 months of age, the indicator is calculated as the proportion of children less than 12-23 months of age receiving one dose of measles-containing vaccine.

MCV2.csv: the percentage of children who have received two doses of measles containing vaccine (MCV2) in a given year, according to the nationally recommended schedule.

PAB.csv: the proportion of neonates in a given year that can be considered as having been protected against tetanus as a result of maternal immunization.

PCV3.csv: the percentage of 1-year-olds who have received three doses of pneumococcal conjugate vaccine (PCV3) in a given year.

Pol3.csv: the percentage of 1-year-olds who have received three doses of polio vaccine in a given year.

ROTAC.csv: the percentage of surviving infants who received the final recommended dose of rotavirus vaccine, which can be either the 2nd or the 3rd dose depending on the vaccine in a given year.


who-immunization-coverage folder

We will take the year 2018 (most recent and less NaN values).

import glob

for files in glob.glob("variable/Global_variable_per-country/who-immunization-coverage/*.csv"):
    data_immuno = pd.read_csv(files)
    print(files)
    data_immuno = data_immuno[[' 2018','Country']]
    df = df.merge(data_immuno, how='left', left_on='Territory', right_on='Country')
    name = files[39:43]
    print(name)
    df[name]= df[' 2018']
    df.drop([' 2018','Country'], axis=1, inplace=True)
    


df.columns

## Split by continent

continent = pd.read_csv('variable/Global_variable_per-country/locations.csv')
continent.drop(['location', 'population_year','population'], axis=1, inplace=True)
df = df.merge(continent, how='left', left_on='Territory', right_on='countriesAndTerritories')
df.drop(['countriesAndTerritories'], axis=1, inplace=True)

index = df[df['continent'].isnull()].index
variable = 'continent'
df.at[index[0], variable] = 'Island'
df.at[index[1], variable] = 'Island'
df.at[index[2], variable] = 'South America'
df.at[index[3], variable] = 'Europe'
df.at[index[4], variable] = 'Asia'
df.at[index[5], variable] = 'Africa'
df.at[index[6], variable] = 'Island'
df.at[index[7], variable] = 'Africa'
df.at[index[8], variable] = 'Africa'
df.at[index[9],variable] = 'Africa'
df.at[index[10],variable] = 'Africa'
df.at[index[11], variable] = 'South America'
df.at[index[12], variable] = 'Africa'
df.at[index[13], variable] = 'Island'
df.at[index[14],variable] = 'Island'
df.at[index[15],variable] = 'South America'
df.at[index[16],variable] = 'Africa'
df.at[index[17], variable] = 'South America'
df.at[index[18], variable] = 'Island'
df.at[index[19], variable] = 'Island'
df.at[index[20], variable] = 'Island'
df.at[index[21], variable] = 'Island'
df.at[index[22],variable] = 'Island'
df.at[index[23],variable] = 'Island'
df.at[index[24],variable] = 'Island'
df.at[index[25],variable] = 'Africa'
df.at[index[26],variable] = 'Asia'
df.at[index[27],variable] = 'Asia'
df.at[index[28],variable] = 'Europe'
df.at[index[29],variable] = 'Island'
df.at[index[30],variable] = 'Island'
df.at[index[31],variable] = 'Island'
df.at[index[32], variable] = 'Africa'
df.at[index[33],variable] = 'Europe'
df.at[index[34],variable] = 'Island'
df.at[index[35], variable] = 'Asia'
df.at[index[36], variable] = 'Asia'
df.at[index[37], variable] = 'Island'
df.at[index[38], variable] = 'Island'
df.at[index[39],variable] = 'Europe'
df.at[index[40], variable] = 'Asia'
df.at[index[41], variable] = 'Africa'
df.at[index[42], variable] = 'Asia'
df.at[index[43], variable] = 'Africa'
df.at[index[44], variable] = 'Africa'
df.at[index[45], variable] = 'South America'
df.at[index[46], variable] = 'Asia'
df.at[index[47], variable] = 'Island'
df.at[index[48], variable] = 'Island'
df.at[index[49],variable] = 'Europe'
df.at[index[50], variable] = 'Island'
df.at[index[51],variable] = 'Europe'
df.at[index[52], variable] = 'North America'
df.at[index[53], variable] = 'Asia'
df.at[index[54], variable] = 'Asia'
df.at[index[55], variable] = 'Island'
df.at[index[56], variable] = 'Asia'
df.at[index[57], variable] = 'Africa'
df.at[index[58], variable] = 'Island'
df.at[index[59], variable] = 'Island'
df.at[index[60], variable] = 'Island'
df.at[index[61], variable] = 'Island'
df.at[index[62], variable] = 'Africa'
df.at[index[63], variable] = 'Island'
df.at[index[64], variable] = 'Island'
df.at[index[65], variable] = 'Africa'
df.at[index[66], variable] = 'Island'
df.at[index[67], variable] = 'Island'
df.at[index[68], variable] = 'Island'
df.at[index[69], variable] = 'Island'
df.at[index[70], variable] = 'Africa'
df.at[index[71], variable] = 'Island'
df.at[index[72], variable] = 'Island'
df.at[index[73], variable] = 'Island'
df.at[index[74], variable] = 'Island'
df.at[index[75], variable] = 'Asia'
df.at[index[76], variable] = 'Island'
df.at[index[77], variable] = 'Island'
df.at[index[78], variable] = 'Africa'
df.at[index[79],variable] = 'Asia'
df.at[index[80], variable] = 'Island'
df.at[index[81],variable] = 'Asia'
df.at[index[82], variable] = 'Island'
df.at[index[83], variable] = 'Island'
df.at[index[84],variable] = 'Asia'

df.continent.unique()

## Commorbidity / Risk Factors

### Obesity 

https://www.kaggle.com/arttua/who-obesity-by-country-2016

obesity = pd.read_csv('variable/Global_variable_per-country/obesity.csv')
obesity['Country'] = obesity['Unnamed: 0']
obesity.drop('Unnamed: 0', axis=1, inplace=True)

obesity.rename(columns={'Both.sexes':'pop_obsesity',
                        'Male':'Male_obesity','Female':'Female_obesity'}, inplace=True)

df = df.merge(obesity, how='left', left_on='Country/Region', right_on='Country')
df.drop(['Country'], axis=1, inplace=True)

## Medical care

### Hospital beds

https://www.kaggle.com/hamzael1/hospital-beds-by-country

beds = pd.read_csv('variable/Global_variable_per-country/beds.csv')

beds.count() / len(beds)

We will take the year 2011.

beds= beds[['Country Code', '2011']]

beds.rename(columns={'2011':'nb_beds'}, inplace=True)

df = df.merge(beds, how='left', left_on='Alpha-3 code', right_on='Country Code')
df.drop(['Country Code'], axis=1, inplace=True)

df.shape

### Health Systems - WHO 


hs = pd.read_csv('variable/Global_variable_per-country/Health_systems.csv')
hs.drop(['Health_exp_pct_GDP_2016','Health_exp_public_pct_2016','Health_exp_out_of_pocket_pct_2016','Health_exp_per_capita_USD_2016','per_capita_exp_PPP_2016','External_health_exp_pct_2016','Completeness_of_birth_reg_2009-18','Completeness_of_death_reg_2008-16'], axis=1, inplace=True)

hs.drop(['Province_State','World_Bank_Name'], axis=1, inplace=True)

hs

hs.head()

hs.Country_Region

df.merge(hs, how='left', left_on='Territory', right_on='Country_Region')

### Pollution ?
https://www.kaggle.com/kweinmeister/pm25-global-air-pollution-20102017

df.shape

df.shape

df.to_csv('Global_variable_per_country.csv',index=False)
