==> zindi_challenge/2.Intervention.ipynb <==
# Pharmaceutical and Non-pharmaceutical interventions

April 13

import pandas as pd
import seaborn as sns
import matplotlib as plt 

#Read train 
deaths = pd.read_csv('data/deaths.csv')
cases = pd.read_csv('data/cases.csv')
deaths.drop(['Province/State'], axis=True)
deaths.shape

df = pd.read_csv('Global_variable_per_country.csv')

df.shape

### Test

https://www.kaggle.com/skylord/covid19-tests-conducted-by-country

test = pd.read_csv('variable/Intervention/test/Tests_conducted_09April2020.csv')

test['As of'].unique()

len(test['Country or region'].unique())

test.drop(['Ref.', 'Source_1','Source_2'], axis=1, inplace=True)

df = df.merge(test, how='left', left_on='Territory', right_on='Country or region')
df.drop(['Country or region'], axis=1, inplace=True)

df[df.Tests.isnull()]

df["Tests"] =  df["Tests"].fillna(0)
df.drop(['As of'], axis=1, inplace=True)



df.shape

247 counties without info relative to test !

Maybe find better data.

### Lockdown 

https://www.kaggle.com/jcyzag/covid19-lockdown-dates-by-country

lock = pd.read_csv('variable/Intervention/lockdown/countryLockdowndates.csv')
lock.drop(['Province', 'Reference'], axis=1, inplace=True)
lock.rename(columns={'Country/Region':'reg','Date':'lockdown_date',
                        'Type':'Type_lockdown'}, inplace=True)
lock =lock.dropna()
lock = lock.drop_duplicates(subset=['reg'], keep='last')

lock.head(1)

df = df.merge(lock, how='left', left_on='Territory', right_on='reg')

df.drop(['Country/Region_y'], axis=1, inplace=True)

df.shape

## Vaccine - clinical test
https://www.clinicaltrials.gov/ct2/results?cond=COVID-19

https://www.who.int/ictrp/en/

https://www.kaggle.com/zohrarezgui/coronavirus-clinical-trials-dataset

trial = pd.read_csv('variable/Intervention/COVID19-927-trials.csv')
trial = trial[['Countries', 'Date registration','Date enrollement','Target size']]
trial.shape

We will take the sum of the number of clinical trials.

df.shape

nb_clinical_trial = list(trial.groupby('Countries').size())
country = list(trial.sort_values('Countries').Countries.unique())[0:62]
trial2 = pd.DataFrame({'nb_trial':nb_clinical_trial ,'country' :country })

def tidy_split(df, column, sep=';', keep=False):
    indexes = list()
    new_values = list()
    df = df.dropna(subset=[column])
    for i, presplit in enumerate(df[column].astype(str)):
        values = presplit.split(sep)
        if keep and len(values) > 1:
            indexes.append(i)
            new_values.append(presplit)
        for value in values:
            indexes.append(i)
            new_values.append(value)
    new_df = df.iloc[indexes, :].copy()
    new_df[column] = new_values
    return new_df

trial2 = tidy_split(trial2, 'country')

nb_clinical_trial = list(trial2.groupby('country').size())
country = list(trial2.sort_values('country').country.unique())
final_trial = pd.DataFrame({'nb_trial':nb_clinical_trial ,'country' :country })
final_trial.head(4)

df.shape

df = df.merge(final_trial, how='left', left_on='Territory', right_on='country')

df["nb_trial"] =  df["nb_trial"].fillna(0)

df.to_csv('Global_variable_per_country+Intervention.csv',index=False)
