==> zindi_challenge/3.Africa.ipynb <==
### Africa

import pandas as pd
import seaborn as sns
import matplotlib as plt 

#Read train 
deaths = pd.read_csv('data/deaths.csv')
cases = pd.read_csv('data/cases.csv')
deaths.drop(['Province/State'], axis=True)
deaths.shape

df = pd.read_csv('Global_variable_per_country+Intervention.csv')
df.shape

### Ebola 

https://www.kaggle.com/imdevskp/ebola-outbreak-20142016-complete-dataset

ebola = pd.read_csv('variable/Africa/ebola_2014_2016_clean.csv')
ebola = ebola[ebola.Date == '2016-03-23']
ebola = ebola[['Country','No. of confirmed deaths', 'No. of confirmed cases']]
ebola.rename(columns={'No. of confirmed deaths':'deaths_ebola',
                        'No. of confirmed cases':'cases_ebola'}, inplace=True)
ebola = ebola.drop([2482,2479], axis=0)

ebola

ebola.loc[2480,['deaths_ebola']] = 3
ebola.loc[2484,['Country']] = 'US'

df = df.merge(ebola, how='left', left_on='Territory', right_on='Country')
df.drop(['country', 'Country'], axis=1, inplace=True)
df["deaths_ebola"] = df["deaths_ebola"].fillna(0)
df["cases_ebola"] = df["cases_ebola"].fillna(0)

df.shape

## Looking at missing values

df.drop(['Alpha-2 code','Country_x','WEBNAME','reg','Country Name','Region'], axis=1, inplace=True)
df.shape

df["Tests/ million"] = df["Tests/ million"].fillna(0)

percent_missing = df.isnull().sum() * 100 / len(df)
missing_value_df = pd.DataFrame({'column_name': df.columns,
                                 'percent_missing': percent_missing})

We will supress all of this variables to predict in a 1st time.

missing_value_df[percent_missing!=0]

df.IncomeGroup

df["nb_beds"] = df["nb_beds"].fillna(0)

### less than 7% missing values

# We assume it's 0 if no records related to medecine in a country.
df["nb_beds"] = df["nb_beds"].fillna(0)
df["HepB3."] = df["HepB3."].fillna(0)
df["DTP3.c"] = df["DTP3.c"].fillna(0)
df["MCV2.c"] = df["MCV2.c"].fillna(0)
df["MCV1.c"] = df["MCV1.c"].fillna(0)
df["Pol3.c"] = df["Pol3.c"].fillna(0)

for lockdown_date and Type_lockdown we will put a category others for unknowm. 

df["lockdown_date"] =  df["lockdown_date"].fillna('Unknown')
df["Type_lockdown"] =  df["Type_lockdown"].fillna('Unknown')

### Less than 1% missing values

na =missing_value_df[(percent_missing<6) & (percent_missing!=0 )].column_name

for var in na:
    print(na)
    df.fillna(df[var].mean(), inplace=True)

df.to_csv('Global_variable_per_country+Intervention+Africa.csv',index=False)
