==> zindi_challenge/4.Cases.ipynb <==
## Final File 

import pandas as pd
import seaborn as sns
import matplotlib as plt 

#Read train 
deaths = pd.read_csv('data/deaths.csv')
cases = pd.read_csv('data/cases.csv')
deaths.drop(['Province/State'], axis=True)
deaths.shape

df = pd.read_csv('Global_variable_per_country+Intervention+Africa.csv')
df.shape

percent_missing = df.isnull().sum() * 100 / len(df)
missing_value_df = pd.DataFrame({'column_name': df.columns,
                                 'percent_missing': percent_missing})

# no missing values
missing_value_df[missing_value_df.percent_missing!=0]

We will add the number of cases. 

cases.drop(['Lat','Long','Province/State','Territory'], axis=1, inplace=True)
cases.shape

cols_cases = cases.columns.tolist()
li = ['c']
for val in cols_cases[1:len(cols_cases)+1] :
    d = val+'cases'
    li.append(d)
cases.columns = li

cases.head()

df.head(1)

cases.shape

df.drop('Province/State', axis=1, inplace=True)

df =pd.concat([df, cases], axis=1)

df.drop(['c'], axis=1, inplace=True)

df.shape

df.to_csv('Final-DATA.csv',index=False)

cols = df.columns.tolist()

cols

percent_missing = df.isnull().sum() * 100 / len(df)
missing_value_df = pd.DataFrame({'column_name': df.columns,
                                 'percent_missing': percent_missing})
missing_value_df
